+++ src/app/app-routing.module.ts
@@ -4,12 +4,14 @@ import { RouterModule, Routes } from '@angular/router';
 
 import { HomeComponent } from './home/home.component';
 import { ImageCropperComponent } from './image-cropper/image-cropper.component';
+import { QuoteGeneratorComponent } from './quote-generator/quote-generator.component';
 import { IpAddressFinderComponent } from './ip-address-finder/ip-address-finder.component';
 
 const routes: Routes = [
-{ path: 'ip-address-finder', component: IpAddressFinderComponent },
-  { path: 'image-cropper', component: ImageCropperComponent },
-  { path: '', component: HomeComponent }
+	{ path: 'ip-address-finder', component: IpAddressFinderComponent },
+	{ path: 'quote-generator', component: QuoteGeneratorComponent },
+	{ path: 'image-cropper', component: ImageCropperComponent },
+	{ path: '', component: HomeComponent }
 ]
 
 @NgModule({