+++ src/app/quote-generator/quote-generator.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-quote-generator',
+  templateUrl: './quote-generator.component.html',
+  styleUrls: ['./quote-generator.component.css']
+})
+export class QuoteGeneratorComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}